fix: renamed files and folders with underscores to be recognized by pytest
