+++ .github/workflows/crm-lambda-ci.yml
@@ -5,7 +5,7 @@ on:
     branches:
       - '**'
     paths:
-      - 'pipelines/crm-interaction/lambda-fn/**'
+      - 'pipelines/crm_interaction/lambda_fn/**'
       - 'requirements.txt'
       - '.github/workflows/**'
     types: [opened, synchronize, reopened]
@@ -24,11 +24,14 @@ jobs:
         with:
           python-version: '3.11'
 
+      - name: Set PYTHONPATH
+        run: echo "PYTHONPATH=$GITHUB_WORKSPACE" >> $GITHUB_ENV
+
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           pip install pytest
-          pip install -r pipelines/crm-interaction/lambda-fn/requirements.txt
+          pip install -r pipelines/crm_interaction/lambda_fn/requirements.txt
 
       - name: Run unit tests
-        run: pytest pipelines/crm-interaction/lambda-fn/tests --disable-warnings -v
+        run: pytest pipelines/crm_interaction/lambda_fn/tests --disable-warnings -v
