+++ pipelines/crm_interaction/lambda_fn/crm_transformation_fn.py
