+++ pipelines/crm_interaction/lambda_fn/crm_validation_fn.py
