+++ pipelines/crm_interaction/lambda_fn/requirements.txt
